Renamed .h files to .hpp for C++ alignment
